+++ saleor/product/models.py
@@ -62,6 +62,7 @@ class Category(ModelWithMetadata, MPTTModel, SeoModel):
     slug = models.SlugField(max_length=255, unique=True, allow_unicode=True)
     description = SanitizedJSONField(blank=True, null=True, sanitizer=clean_editor_js)
     description_plaintext = TextField(blank=True)
+    updated_at = models.DateTimeField(auto_now=True, blank=True, null=True)
     parent = models.ForeignKey(
         "self", null=True, blank=True, related_name="children", on_delete=models.CASCADE
     )
@@ -82,6 +83,7 @@ class Category(ModelWithMetadata, MPTTModel, SeoModel):
                 fields=["name", "slug", "description_plaintext"],
                 opclasses=["gin_trgm_ops"] * 3,
             ),
+            BTreeIndex(fields=["updated_at"], name="updated_at_idx"),
         ]
 
     def __str__(self) -> str: